Fix client notify controller on command success
The reason for the previous noticed issue was that the controller
was simply not being notified that the connection was "good" again.
This is a problem with only establishing a new connection on
ECONNRESET/EPIPE, which stems from not having any control of the
"server", in this case a projector.  I can't force it to realize when
the "client" has disconnected, without properly shutting down the
socket, which can't be done without a working connection to the device.
If the cable gets disconnected, we don't want to keep spamming and
filling up our log with useless "command timed out" messages.

So we settle for a boolean Client.lastCommandTimedOut that if true
causes the client to fire a fake "ClientConnected" event once it appears
that we once again are able to send commands and receive responses.