Add solution, break into separate projects
Broke everything into separate projects to attempt to make it easier
to reuse projector classes and other things later.  Also had to
eliminate a circular dependency in the NEC projector classes
(dependency on MainWindow).  Probably the right thing.  Didn't really
know where I was going with that.  I will need a way for the controller
classes to notify the app/window when it should update its visuals
to indicate things happening, but I'll cross that bridge at a later time.