Add partially working VolumeUp/Down
VolumeUp works, VolumeDown not so much.  It's also not very useful for a number of reasons, so I may end up scrapping these methods completely.  The projector I'm testing with doesn't visually indicate that the volume increased after calling VolumeUp, but thumbing thru the menu confirms it.  Plus who the hell is going to use the builtin projector speakers?