Separate IDisplay from IInputSelectable
Took SelectInput out of IDisplay and put it in its own interface.  Also moved PowerOnSelectInput specifically into the Projector & TV abstract classes as it kinda doesn't make sense for it to be in either the IDisplay or IInputSelectable interfaces... since it does the work of both.  Do I need an interface that combines the two and is implemented by Projector & TV?  Geez, when is enough enough?  Moving a little more toward explicitly implemented interfaces, but I still like having a Projector / TV catch-all that just scoops up a bunch of interfaces and lets you do real work without type casts and null checking.