Add GetStatus(bool appWantsText)
This is a hideous hack that will be used by the DisplayTester app for showing a big block of debugging text in its TextView.  Since it's really only for the DisplayTester, I put the method in the IDisplay interface.  I'm not at all happy with this solution, but we'll keep it for now and keep working toward something better.