Refactor device/projector connection system
I might one day decide to implement connecting to devices over a serial port, so I've refactored a little bit to allow that to happen.

We now use an interface to encapsulate connection information and pass that to classes that might be able to choose to connect using either network or serial ports.  Devices that can connect either way (like the NEC projectors) will work with their "client" objects through interfaces.  Required lots of small changes in lots of places.